+++ hypervisor/acpi_parser/acpi_ext.c
@@ -30,6 +30,7 @@
 #include <rtl.h>
 #include "acpi.h"
 #include <x86/pgtable.h>
+#include <x86/apicreg.h>
 #include <x86/ioapic.h>
 #include <logmsg.h>
 #include <x86/host_pm.h>
