+++ hypervisor/include/arch/x86/apicreg.h
@@ -29,8 +29,6 @@
 #ifndef APICREG_H
 #define APICREG_H
 
-#include <x86/page.h>
-
 /*
  * Local && I/O APIC definitions.
  */
